SUV computation for PET images with START decay correction is now done using .
Earlier it was assumed that the time to which images were decay corrected was the Acquisition Time (see DICOM C.8.9.1.1.5). However, this is not required by documentation of the QIBA SUV technical subcommittee (https://qibawiki.rsna.org/index.php/Standardized_Uptake_Value_(SUV)). To not rely on Series Time and Series Date DICOM attributes, the back-computation solution was used.