WIP on self-segmenting structure sets.
I.e. structure sets that do not draw upon an associated image, but rather utilises internal information for generating the segmentation mask.

This is helpful in case the RTSTRUCT set shares the frame of reference with an image, but a one-to-one mapping of image slices and contours is not possible.